fix background hero static link to dynamic link
