+++ .env.example
@@ -1,3 +1,5 @@
+BASE_URL=root_url
+
 DB_HOST=host
 DB_NAME=name_database
 DB_USER=username
