+++ composer.json
@@ -1,6 +1,7 @@
 {
   "require": {
     "phpmailer/phpmailer": "^6.6",
-    "dompdf/dompdf": "^2.0"
+    "dompdf/dompdf": "^2.0",
+    "vlucas/phpdotenv": "^5.5"
   }
 }
