+++ package.json
@@ -1,8 +1,7 @@
 {
   "devDependencies": {
+    "flowbite": "^1.5.4",
     "tailwindcss": "^3.1.8"
   },
-  "dependencies": {
-    "flowbite": "^1.5.3"
-  }
+  "dependencies": {}
 }
